BUG: locale dropdown sometimes hidden after panel reload
The locale dropdown was sometimes hidden after a panel reload via ajax,
due to the caching behaviour for the panel. This was caused by the fact
that the dropdown was hidden at the point in which the panel was cached
(probably due to chosen).
Fixed by always making sure the dropdown is shown on its onmatch function.